Refactor RentalInfo into Service.
